+++ CHANGELOG.md
@@ -219,6 +219,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Added `XLACheckpointIO` plugin ([#9972](https://github.com/PyTorchLightning/pytorch-lightning/pull/9972))
 
+- Added `ckpt_path` argument for `trainer.fit()` to replace `Trainer(resume_from_checkpoint=)`) ([#10061](https://github.com/PyTorchLightning/pytorch-lightning/pull/10061))
+
 
 ### Changed